project_starter/requirements.txt
flask

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

google-cloud-resource-manager
